Group form components into one folder
